Wire in a route to retrieve projects.
Projects (for which models are already turned into JSON objects) are now
returned in a JSON array, rather than as a newline separated string of
JSON objects.  This means that it is trivial for the client to
deserialise the entire payload, without having to first perform custom
parsing to split the string into a collection of JSON objects that can
each then be deserialised.