Use a URI-friendly string for resource IDs.
Projects, methods and samples lend themselves to being represented as
resources in a RESTful API.  But, although the names of our resources
are unique, they are not used as the primary identifier.  Rather we use
hashids, which leaves open the possiblity of renaming a resource without
breaking clients that may have cached the name as a resource identifier;
cf. http://hashids.org.

Thus you can do something like:

   curl http://sagittariidae/projects/xY4fj-inconel/samples/tJ3rP-sample-123

rather than:

   curl "http://sagittariidae/sample?project-id=123&sample-id=456"

with the former providing resource labels in addition to the IDs and an
explicitly hierarchical representation.