Move JSONising out of `models` module.
That the API returns data as JSON documents is a presentation issue and
not something about which the persistance layer should be concerned.
This also makes it easier to re-use the functions that retrieve model
instances from within the module.

While I'm not wild about leaking the db model classes into the
presentation layer, it does make parts of the serialisation easier (like
the custom project name serialization for samples), but at the cost of
tighter coupling between the presentation layer and the implementation
of the persistance framework (SQLAlchemy).

But unfortunately, while it is easy to extend Python's json module to
new types, it is hard to extend for built-in types.  So using dicts
instead of the models to communicate between the API and persistance
layers isn't easily possible (and would require way more monkey-patching
than I would consider reasonable).  The alternative is to create a class
hierarchy that parallels/wraps the Model classes used by the persistance
layer.

At this point though that seems like overkill.