Retrieve a project sample with its test stages.
The stages are massaged somewhat to get a suitable JSON representation;
the ordinal value of each stage's position in the sequence is attached
to the obfuscated ID in order to provide some context.  The stage ID is
not used since it is global across all samples and thus is it likely
that sequence of stage IDs for any one sample will have gaps, making
them meaningless as an external identifier.

Retrieving a specific stage will ultimately be a GET of something like:
  http://sagittariidae/projects/<p>/samples/<s>/stages/EYore-9