Wire in routes to GET and PUT sample test stages.
Sample stages must be put using a token.  Attempting to add a stage with
in an invalid token will result in a 409 (Conflict).  The token reflects
the state of the database at the time of a `get_sample_stages` call and
must match the state of the DB at the time of the `add_sample_stage`
call for it to be successful.

This puts an additional burden on the client to ensure that its view of
the stages is up to date and risks starvation if the mutation rate of
the stages is high.  This is still better, however, than having spurious
data being written to the DB.  Fortunately, the expected rate of change
of change is low, so in practice neither starvation nor the need for
frequent updates should be problematic.

To avoid duplication stages are no longer included in JSON serialised
sample responses.  In general, we should favour NOT including
collections in higher-level entities, and rather including parent
context in lower-level entities (e.g. a sample won't include its stages,
but a stages response will include the sample as context).