Wire up file upload routes.
Two new API routes have been added: one to upload a file in parts, and
another to signal the completion of upload and reconstitute the file
from its parts.

This initial commit of this functionality does reconstitute the file,
but doesn't yet allocate it to the appropriate sample stage.  It is
anticipated that these uploaded files will be archives containing
the many files produced by a scan or analysis stage, and these will need
to be unpacked and put into an appropriate location.

It is also because of this additional server-side processing that we
respond to a successful file upload completion request with a 202,
rather than the default 200.  The 202 response code indicates that all
client interaction has been successfully handled, but that further
server-side processing is required before the resource will become
available.