Wire up file upload routes.
Two new API routes have been added: one to upload a file in parts, and
another to signal the completion of upload and reconstitute the file
from its parts.

This initial commit of this functionality does reconstitute the file,
but doesn't yet allocate it to the appropriate sample stage.  It is
anticipated that these uploaded files will be archives containing
the many files produced by a scan or analysis stage, and these will need
to be unpacked and put into an appropriate location.