Create models for user auth{entica,oriza}tion.
The intention is for Sagittariidae not to be the primary agent for user
authentication.  Rather, the responsibility for authentication is
delegated to the client (presumably Google or Facebook authentication
via something like Firebase).  Ultimately the WS should receive a user
token and provider info, and validate the validaty of the token with the
provider.

While the authentication is handled externally, the authorization is
managed by Sagittariidae.  Since we're not committing to Firebase for
storage, we need to track users that have been granted access to the
system.

The downside to leaving the authentication process up to the client is
that it makes it harder to use the web service from a command-line
client; it is not clear to me at this point what tools are available for
generating user tokens outside of a browser environment (where
Firebase's `signInWithRedirect` won't function).

This approach also requires that communication between client and server
is secured with SSL!  Otherwise we risk exposing user tokens to
eavesdroppers and opening ourselves up to replay attacks.