Fix broken JSON model decoding.
The mechanism to resolve the decode function for a particular type of
`Model` is broken.  Looking specifically at the `__class__` attribute
limits resolution only to that class and non of its subclasses.

Thus the generic decoding for any subclass of db.Model would never have
been triggered.

Also, the validation of the model attributes `id` and `obfuscated-id` is
no longer required since `_dictify` doesn't operate on these attributes
any more.