Wire in routes to add/get user info.
This commit sees in the introduction of API calls requiring user
authentication.  Authentication is assumed to have been handled by
third-party providers (e.g. Google) whose auth tokens can be
independently verified.

Authenticated users are allowed to request access to the system by
adding themselves as new users.  These requests are acknowledged with a
'202' to indicate that the request has been accepted, but may not yet
have been fully processed.  In our case, the new user record is marked
as 'pending' and requires an out-of-band change for the user to be made
active.

User authentication and authorization must be retro-fitted to existing
API functions.  This will follow in future commits.