Require authorization for certain user actions.
Authorization is achienved by wrapping the action to be performed in a
`with user_authorization` block.  While is works, it not an ideal model
as it is possible for conflicting, or no, authorization rules to be
applied for any given object depending on how the actions are wrapped.

A better model would be to apply the permissions to the objects; then,
regardless of how the object is obtained, or what actions are being
performed, a consistent set of rules can be applied.