Add mechanism to 'complete' uploaded files.
After having been uploaded, sample stage files are moved into place
asynchronously.  This allows us to more easily keep the filesystem and
the database in sync in the face of unexpected failures.

Should the service crash, or be unexpectedly terminated before the file
has been moved into place, the asynchronous sweeper can retry the
operation and drive synchronization between the two stores.

Note that when files are added to the store, they are added in a
'staged' state, which indicates that any upload processing has been
completed and that the files are ready to be moved into place.  This
intentionally decouples the upload processing from the process of
associating a file with a stage.  The distinction will allow us to
accomodate a 'batch mode' upload in future, where an archive containing
multiple files is uploaded, but it is the individual files that are
associated with the stage.