Deployment preparation.
* Serve static site data.

Configure Werkzeug to serve the site files: static HTML and
{Clojure,Java}Script.  The 'static' directory from which there are
served are not part of this package, but will be created during
deployment.

* Change database location.

The database has been moved to a location that is more consistent with
the Linux FS standard.  Separate from the project source, persistance
can be achieved by using the directory as a mount point for a Docker
volume.