Preserve extensions in unique datafile names.
To preserve the immutability of datafiles, we attempt to version them by
appending a "version" counter as a suffix.  This makes the type of the
file unrecognizable when downloaded to systems that depend on the
extension to determine the type.

Rather than simply appending the counter value, we now attempt to insert
it into the file name in a smart way that doesn't break extension-based
type detection.