Exclude container "extension" from ext detection.
Some methods produce a collection of files that are only useful as a
collection, but the formats in which they are produced may not provide
their own container format (as TIFF does, for example).

To avoid storing 10s (or even 100s) of discrete files when a container
format isn't available (or used), we define our own container format.
Files of this type have the custom extension that starts with
"container-".  The suffix is a content label that will typically be the
extension of the individual files, e.g. "container-dicom".

Files with this custom extension must always be excluded from the
extension detection, so that the counter can be inserted correctly.